move verify_args_struct_by_command_run_with_no_genesis_fetch()
